restructured the files to make it more organized. added an input for the edit functionality
