+++ src/components/HomePage/HomePageGrid.jsx
@@ -6,11 +6,11 @@ import { Grid, Image, Input, Menu } from "semantic-ui-react";
 import { Link } from 'react-router-dom'
 
 // image imports 
-import book from "./images/joke-book.png"; 
-import feed from "./images/public-feed.png"; 
-import favorites from "./images/favorites.png"; 
-import friends from "./images/friends.png"; 
-import profile from "./images/profile.png"; 
+import book from "../images/joke-book.png"; 
+import feed from "../images/public-feed.png"; 
+import favorites from "../images/favorites.png"; 
+import friends from "../images/friends.png"; 
+import profile from "../images/profile.png"; 
 // import inbox from "./images/inbox.png"; 
 
 const HomePageGrid = (props) => {
