+++ src/components/PrivateJokes/JokeList.js
@@ -1,9 +1,9 @@
 import React, { useState, useEffect}  from "react";
 import JokeCard from "./JokeCard";
-import { jokeContainer, JokeCarder } from './StyledWidgets.js';
+import { jokeContainer, JokeCarder } from '../../styles/StyledWidgets';
 import Jokes from './Jokes'
 import { connect } from 'react-redux'
-import { fetchData, addData, deleteData, editData } from '../store/actions'
+import { fetchData, addData, deleteData, editData } from '../../store/actions'
 import { Button } from "semantic-ui-react";
 
 function JokeList (props) {
