+++ src/components/PrivateJokes/Jokes.js
@@ -1,12 +1,12 @@
 import React, { useState, useEffect } from "react";
 import { connect } from 'react-redux';
-import { fetchData, addData, deleteData, editData } from '../store/actions';
+import { fetchData, addData, deleteData, editData } from '../../store/actions';
 import JokeList from './JokeList';
-import TopHalfPage from "./TopHalfPage";
-import Footer from "./Footer";
+import TopHalfPage from "../HomePage/TopHalfPage";
+import Footer from "../NavFooter/Footer";
 import styled from "styled-components";
-import {JokeListHeading} from './StyledWidgets';
-import AddJokeCircle from "./images/add-img-circle.png"
+import {JokeListHeading} from '../../styles/StyledWidgets';
+import AddJokeCircle from "../images/add-img-circle.png"
 
 
 const Title = styled.h1 `
