+++ src/styles/StyledWidgets.js
